# backend/api/swagger.yaml
openapi: 3.0.0
info:
  title: ParticipantCaller API
  description: API for the app ParticipantCaller
  contact:
    email: aiqueneldar@gmail.com
  license:
    name: GPL 3.0
    url: https://www.gnu.org/licenses/gpl-3.0.en.html
  version: 1.0.0
servers:
- url: https://virtserver.swaggerhub.com/p9139/partcaller/1.0.0
  description: ParticipantCaller Enpoint
...
paths:
...
  /entities/list:
    get:
      tags:
      - entities
      summary: List all entities UUIDs
      description: |
        Returns a list of all enteties UUIDs and names
      operationId: listEntities
      responses:
        "200":
          description: list of entities with summary information
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EntityIdName'
                x-content-type: application/json
        "400":
          description: "An invalid input, or object invalid"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invalid'
        "401":
          description: Client is not logged in and tries to reach a resource that
            needs authorization
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        "403":
          description: Client is actorhized but does not have authorization to use
            this resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Forbidden'
        "500":
          description: An internal server error occurred
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /entities/get:
    get:
      tags:
      - entities
      summary: Retrive entity information
      description: "Use  with query string 'id' containing the entity id to  \nget\
        \ specific information about a entity.\n"
      operationId: getEntities
      parameters:
      - name: id
        in: query
        description: |
          UUID of the entity to be retreived
        required: true
        style: form
        explode: true
        schema:
          type: string
      responses:
        "200":
          description: Detailed description of one entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Entity'
          links:
            ModifyEntity:
              $ref: '#/components/links/ModifyEntity'
            DeleteEntity:
              $ref: '#/components/links/DelEntity'
        "400":
          description: "An invalid input, or object invalid"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invalid'
        "401":
          description: Client is not logged in and tries to reach a resource that
            needs authorization
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        "403":
          description: Client is actorhized but does not have authorization to use
            this resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Forbidden'
        "404":
          description: The specified resource was not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
        "500":
          description: An internal server error occurred
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /entities/add:
    post:
      tags:
      - entities
      summary: Adds a new entity
      description: Adds an new entity to the system
      operationId: addEntity
      requestBody:
        description: Information about the new entity to add.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EntityNew'
      responses:
        "201":
          description: A new entity was added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EntityId'
          links:
            GetEntity:
              $ref: '#/components/links/GetEntity'
            ModifyEntity:
              $ref: '#/components/links/ModifyEntity'
            DeleteEntity:
              $ref: '#/components/links/DelEntity'
        "400":
          description: "An invalid input, or object invalid"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invalid'
        "401":
          description: Client is not logged in and tries to reach a resource that
            needs authorization
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        "403":
          description: Client is actorhized but does not have authorization to use
            this resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Forbidden'
        "409":
          description: A duplicate resource already exists. Try using resource/modify
            instead
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Duplicate'
        "500":
          description: An internal server error occurred
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /entities/modify:
    put:
      tags:
      - entities
      summary: Modify an existing entity
      description: Modifies an already existing entity
      operationId: modifyEntity
      requestBody:
        description: "Details to modify about an entity, UUID is immutable"
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Entity'
      responses:
        "201":
          description: Entity is modified
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EntityId'
          links:
            GetEntity:
              $ref: '#/components/links/GetEntity'
            DeleteEntity:
              $ref: '#/components/links/DelEntity'
        "400":
          description: "An invalid input, or object invalid"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invalid'
        "401":
          description: Client is not logged in and tries to reach a resource that
            needs authorization
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        "403":
          description: Client is actorhized but does not have authorization to use
            this resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Forbidden'
        "404":
          description: The specified resource was not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
        "500":
          description: An internal server error occurred
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /entities/delete:
    delete:
      tags:
      - entities
      summary: Delete an entity
      description: Deletes an already existing entity
      operationId: deleteEntity
      parameters:
      - name: id
        in: query
        description: UUID of the entity to delete
        required: true
        style: form
        explode: true
        schema:
          type: string
      responses:
        "204":
          description: Entity was successfully deleted
        "400":
          description: "An invalid input, or object invalid"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invalid'
        "401":
          description: Client is not logged in and tries to reach a resource that
            needs authorization
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        "403":
          description: Client is actorhized but does not have authorization to use
            this resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Forbidden'
        "404":
          description: The specified resource was not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
        "500":
          description: An internal server error occurred
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
...
    Entity:
      required:
      - id
      - name
      - roles
      type: object
      properties:
        id:
          type: string
          description: UUID for the entity
          example: 5691f5a1-cde4-4b09-8a1a-38fa5ba4adf2
        name:
          type: string
          description: Human readable name of the entity
          example: Player 1
        roles:
          type: array
          items:
            type: string
            description: UUIDs of abilites grouped togheter in this role
            example: 1c546582-0fda-4875-8b26-8dd70599e6ba
        parents:
          type: array
          items:
            type: string
            description: UUIDs of other enteties that are parents to this entity
            example: 5691f5a1-cde4-4b09-8a1a-38fa5ba4adf2
        children:
          type: array
          items:
            type: string
            description: UUIDs of other enteties that are children to this entity
            example: 5691f5a1-cde4-4b09-8a1a-38fa5ba4adf2
        attributes:
          type: array
          description: List of objects that have attributes associated with the entity
          items:
            type: object
            description: An object holding information about the entity
            example:
              JWTtoken: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c
      example:
        children:
        - 5691f5a1-cde4-4b09-8a1a-38fa5ba4adf2
        - 5691f5a1-cde4-4b09-8a1a-38fa5ba4adf2
        roles:
        - 1c546582-0fda-4875-8b26-8dd70599e6ba
        - 1c546582-0fda-4875-8b26-8dd70599e6ba
        name: Player 1
        attributes:
        - JWTtoken: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c
        - JWTtoken: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c
        id: 5691f5a1-cde4-4b09-8a1a-38fa5ba4adf2
        parents:
        - 5691f5a1-cde4-4b09-8a1a-38fa5ba4adf2
        - 5691f5a1-cde4-4b09-8a1a-38fa5ba4adf2
    EntityNew:
      required:
      - name
      type: object
      properties:
        name:
          type: string
          description: Human readable name of the entity
          example: Player 1
        roles:
          type: array
          items:
            type: string
            description: UUIDs of abilites grouped togheter in this role
            example: 1c546582-0fda-4875-8b26-8dd70599e6ba
        parents:
          type: array
          items:
            type: string
            description: UUIDs of other enteties that are parents to this entity
            example: 5691f5a1-cde4-4b09-8a1a-38fa5ba4adf2
        children:
          type: array
          items:
            type: string
            description: UUIDs of other enteties that are children to this entity
            example: 5691f5a1-cde4-4b09-8a1a-38fa5ba4adf2
        attributes:
          type: array
          description: List of objects that have attributes associated with the entity
          items:
            type: object
            description: An object holding information about the entity
            example:
              JWTtoken: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c
    EntityIdName:
      required:
      - id
      - name
      type: object
      properties:
        id:
          type: string
          description: UUID for the entity
          example: 5691f5a1-cde4-4b09-8a1a-38fa5ba4adf2
        name:
          type: string
          description: Human readable name of the entity
          example: Player 1
      example:
        name: Player 1
        id: 5691f5a1-cde4-4b09-8a1a-38fa5ba4adf2
    EntityId:
      required:
      - id
      type: object
      properties:
        id:
          type: string
          description: UUID for the entity
          example: 5691f5a1-cde4-4b09-8a1a-38fa5ba4adf2
      example:
        id: 5691f5a1-cde4-4b09-8a1a-38fa5ba4adf2
...
    Error:
      type: object
      properties:
        errorid:
          type: integer
          description: Error number of the error that occurred. HTTP 5xx code
          format: int32
          example: 500
        errormsg:
          type: string
          description: Error message for the error that occurred
          example: Internal Server Error
    NotFound:
      type: object
      properties:
        errorid:
          type: integer
          description: Error number of the error that occurred. HTTP 4xx code
          format: int32
          example: 404
        errormsg:
          type: string
          description: Error message for the error that occurred
          example: Not Found
    Unauthorized:
      type: object
      properties:
        errorid:
          type: integer
          description: Error number of the error that occurred. HTTP 4xx code
          format: int32
          example: 401
        errormsg:
          type: string
          description: Error message for the error that occurred
          example: Unauthorized
    Forbidden:
      type: object
      properties:
        errorid:
          type: integer
          description: Error number of the error that occurred. HTTP 4xx code
          format: int32
          example: 403
        errormsg:
          type: string
          description: Error message for the error that occurred
          example: Forbidden
    Invalid:
      type: object
      properties:
        errorid:
          type: integer
          description: Error number of the error that occurred. HTTP 4xx code
          format: int32
          example: 400
        errormsg:
          type: string
          description: Error message for the error that occurred
          example: Invalid Object or resource
    Duplicate:
      type: object
      properties:
        errorid:
          type: integer
          description: Error number of the error that occurred. HTTP 4xx code
          format: int32
          example: 409
        errormsg:
          type: string
          description: Error message for the error that occurred
          example: Duplicate object exsists
...
  links:
...
    GetEntity:
      operationId: getEntity
      parameters:
        id: $response.body#/id
      description: Get entity full information
    ModifyEntity:
      operationId: modifyEntity
      parameters:
        id: $response.body#/id
      description: "Modify entity, modify any information beside UUID which is immutable."
    DelEntity:
      operationId: deleteEntity
      parameters:
        id: $response.body#/id
      description: Delete entity
...