# backend/go/model_entity_id_name.go
/*
 * ParticipantCaller API
 *
 * API for the app ParticipantCaller
 *
 * API version: 1.0.0
 * Contact: aiqueneldar@gmail.com
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package swagger

type EntityIdName struct {
	// UUID for the entity
	Id string `json:"id"`
	// Human readable name of the entity
	Name string `json:"name"`
}