# backend/go/model_entity_new.go
/*
 * ParticipantCaller API
 *
 * API for the app ParticipantCaller
 *
 * API version: 1.0.0
 * Contact: aiqueneldar@gmail.com
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package swagger

type EntityNew struct {
	// Human readable name of the entity
	Name string `json:"name"`

	Roles []string `json:"roles,omitempty"`

	Parents []string `json:"parents,omitempty"`

	Children []string `json:"children,omitempty"`
	// List of objects that have attributes associated with the entity
	Attributes []interface{} `json:"attributes,omitempty"`
}